+++ CHANGELOG.md
@@ -3,6 +3,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Removed
+- deprecated `\Smsapi\Client\Feature\Contacts\Bag\CreateContactBag::withPhone`
+- deprecated `\Smsapi\Client\Feature\Sms\Bag\ScheduleSmsBag::setIdx`
+- deprecated `\Smsapi\Client\Feature\Sms\Bag\SendSmsToGroupBag::setIdx`
+- deprecated `\Smsapi\Client\Feature\Sms\Bag\ScheduleSmsToGroupBag::setIdx`
+- deprecated `\Smsapi\Client\Curl\Exception\RequestException`
+- deprecated `\Smsapi\Client\Curl\Exception\NetworkException`
+- deprecated `\Smsapi\Client\Curl\Exception\ClientException`
+
 ## [3.0.13] - 2026-01-12
 ### Added
 - PHP-8.5 support